requirements.txt
# BlueKing SaaS core framework
Django>=3.2,<4.0
blueapps[opentelemetry]>=4.15.0

# REST API
djangorestframework>=3.14.0
django-filter>=22.1

# BlueKing IAM (permission center)
bk-iam>=2.0.0

# GitLab API client
python-gitlab>=3.15.0

# HTTP requests
requests>=2.31.0

# WSGI server
gunicorn>=21.2.0

# Database
mysqlclient>=2.2.0

# CORS
django-cors-headers>=4.3.0

# Utilities
python-dotenv>=1.0.0
